Move cache support to Bundle
